+++ src/org/meta_environment/rascal/library/experiments/VL/Shape.java
@@ -13,7 +13,7 @@ public class Shape extends Compose {
 	}
 	
 	@Override
-	void bbox(float left, float top){
+	void bbox(int left, int top){
 		this.left = left;
 		this.top = top;
 		width = 0;
